Fixed my resave problem : issue was with the Controller implementation
Initially I used a Piped Input and Output Streams. However in my initial
implementation all the bytes writing then reading would happen sequentially
in the same thread. This causes issue because the piped stream do have a
buffer capacity. It resulted in Jackson writing too many bytes and then
blocking, waiting for the buffer to be emptied, which could never happen.

Fixed it by replacing this with limitless buffer; but I want to try again
with the read and write happening in different threads just for fun.