Looking at it, the UX is not that great and needs to be reworked thoroughly.
* The single-screen approach leads to a confusing interface.
* Large items are cute but take more spaces, thus more scrolling is
required to access the desired information.
* Some existing features are useless while the features we actually need
are absent.
* To a less problematic degree, the wizard algorithm is correct but
imprecise.

On the other hand this revision is functional and can be actually used,
(ven though it is not yet packaged and can run only using development
tools)

For these reasons I declare this revision as version 1 and tagged as
such, and from there we start a complete redesign of the interface.