Bugfix : when creating a movie, then planning it immediately in a session, the wizard does not crash anymore.
The root cause is that the movies flux consumed by the wizard was not
updated when it should be. Indeed, the movies flux returned only movies planned in a
session, not all movies. So, when a movie is just created, it was not
streamed to the wizard (because not planned yet). Later on, once the
movie is created, the updated list of sessions including the new one is
streamed to the wizard - however, since the movies per se were not
updated, then the movie stream remained the same, ie not including the
new movie.

The result was then that the wizard tried to gather data for a movie
referenced by a session, but which did not exist in the not-updated
movies stream, resulting eventually in an NPE.